Finished reading input

+++ codigo/Courier.cpp
@@ -1,4 +1,4 @@
 #include "Courier.h"
 
-Courier::Courier(const std::string &licensePlate, double volMax, double pesoMax, double transportFee) : licensePlate(
+Courier::Courier(const std::string &licensePlate, unsigned int volMax, unsigned int pesoMax, unsigned int transportFee) : licensePlate(
         licensePlate), volMax(volMax), pesoMax(pesoMax), transportFee(transportFee) {}
